Updates to get this compile under Linux (mostly fixing uint32_t type errors)
One of the issues I kept running into was that I was getting undefined uint32_t (or uint8_t or int8 ... ). My solution (and probably not the correct one) was to modify the espmissingincludes.h with a lot of 'typedef unsigned char uint8;' like lines. I'm not sure why the c_types.h isn't taking care of this. I'll need to look into this. Also I have to grab esptool-ck (https://github.com/igrr/esptool-ck) as the esptool.py and esptool2 wouldn't work directly.